\name{CryptRndTest-internal}

\alias{KSADdga}
\alias{KSADCHRY}
\alias{Random.walk.Y}
\alias{Random.walk.D}
\alias{Random.walk.G}
\alias{dogumGunuAraliklari}
\alias{adaptive.chi.square.default}
\alias{adaptive.chi.square.main}
\alias{birthday.spacings.default}
\alias{birthday.spacings.main}
\alias{book.stack.default}
\alias{book.stack.main}
\alias{GCD.test.default}
\alias{GCD.test.main}
\alias{random.walk.tests.default}
\alias{random.walk.tests.main}
\alias{topological.binary.default}
\alias{topological.binary.main}
\alias{check}

\title{
Functions for internal use only
}
\description{
Contains functions designed for internal use only.
}

\author{
Haydar Demirhan

Maintainer: Haydar Demirhan <haydarde@hacettepe.edu.tr>
}

\seealso{
\code{\link[CryptRndTest]{adaptive.chi.square}}, \code{\link[CryptRndTest]{birthday.spacings}}, \code{\link[CryptRndTest]{book.stack}}, \code{\link[CryptRndTest]{GCD.test}}, \code{\link[CryptRndTest]{GCD}},
\code{\link[CryptRndTest]{GCD.q}},
\code{\link[CryptRndTest]{GCD.big}},
 \code{\link[CryptRndTest]{random.walk.tests}}, \code{\link[CryptRndTest]{topological.binary}}, \code{\link[CryptRndTest]{Strlng2}}, \code{\link[copula]{Stirling2}}
}

