\name{print.CryptRndTest}
\alias{print.CryptRndTest}
\title{
Print Test Results
}
\description{
Prints a summary of test results.
}
\usage{
\method{print}{CryptRndTest}(x,...)
}
\arguments{
  \item{x}{
an object including information to be printed.
}
  \item{...}{
other arguments.
}
}

\author{
Haydar Demirhan

Maintainer: Haydar Demirhan <haydarde@hacettepe.edu.tr>

}