\name{CryptRndTest-internal}

\alias{KSADdga}
\alias{KSADCHRY}
\alias{Random.walk.Y}
\alias{Random.walk.D}
\alias{Random.walk.G}
\alias{dogumGunuAraliklari}

\title{
Functions for internal use only
}
\description{
Contains functions designed for internal use only.
}

\author{
Haydar Demirhan

Maintainer: Haydar Demirhan <haydarde@hacettepe.edu.tr>
}

\seealso{
\code{\link[CryptRndTest]{adaptive.chi.square}}, \code{\link[CryptRndTest]{birthday.spacings}}, \code{\link[CryptRndTest]{book.stack}}, \code{\link[CryptRndTest]{GCD.test}}, \code{\link[CryptRndTest]{GCD}},
\code{\link[CryptRndTest]{GCD.q}},
\code{\link[CryptRndTest]{GCD.big}},
 \code{\link[CryptRndTest]{random.walk.tests}}, \code{\link[CryptRndTest]{topological.binary}}, \code{\link[CryptRndTest]{Strlng2}}, \code{\link[copula]{Stirling2}}
}

